(people query): Implement people query extract
- create a people query component

[Starts #177933454]
